Io corrected

+++ OOPitch.py
@@ -710,8 +710,8 @@ class Match:
                     # objs, = ax.plot(team['ball_x'], team['ball_y'], 'ko', MarkerSize=6, alpha=1.0, LineWidth=0)
                     figobjs.append(objs)
                 # include match time at the top
-                frame_minute = int(frame / (60 * 25))
-                frame_second = np.int_( np.floor((frame / (60 * 25) - frame_minute) * 60.))
+                frame_minute = np.int(frame / (60 * 25))
+                frame_second = np.int( np.floor((frame / (60 * 25) - frame_minute) * 60.))
                 timestring = f"{frame_minute}:{frame_second}"
                 objs = ax.text(-2.5, field_dimen[1] / 2. + 1., timestring, fontsize=14)
                 figobjs.append(objs)
